+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Fixed** for any bug fixes.
 - **Security** in case of vulnerabilities.
 
+## [0.5.0] - 2024-01-14
+
+### Added
+
+- AuthError custom errors
+
 ## [0.4.0] - 2024-01-14
 
 ### Added